feat!: update `github-releases` datasource digest computation to use git tag, and preserve existing digest semantics as separate datasource (#20178)
