+++ lib/modules/datasource/github-tags/index.ts
@@ -2,6 +2,7 @@ import is from '@sindresorhus/is';
 import { logger } from '../../../logger';
 import { queryReleases, queryTags } from '../../../util/github/graphql';
 import type { GithubReleaseItem } from '../../../util/github/graphql/types';
+import { findCommitOfTag } from '../../../util/github/tags';
 import { getApiBaseUrl, getSourceUrl } from '../../../util/github/url';
 import { GithubHttp } from '../../../util/http/github';
 import { Datasource } from '../datasource';
@@ -22,42 +23,6 @@ export class GithubTagsDatasource extends Datasource {
   constructor() {
     super(GithubTagsDatasource.id);
     this.http = new GithubHttp(GithubTagsDatasource.id);
-  }
-
-  async getTagCommit(
-    registryUrl: string | undefined,
-    packageName: string,
-    tag: string
-  ): Promise<string | null> {
-    logger.trace(`github-tags.getTagCommit(${packageName}, ${tag})`);
-    try {
-      const tags = await queryTags({ packageName, registryUrl }, this.http);
-      // istanbul ignore if
-      if (!tags.length) {
-        logger.debug(
-          `github-tags.getTagCommit(): No tags found for ${packageName}`
-        );
-      }
-      const tagItem = tags.find(({ version }) => version === tag);
-      if (tagItem) {
-        if (tagItem.hash) {
-          return tagItem.hash;
-        }
-        logger.debug(
-          `github-tags.getTagCommit(): Tag ${tag} has no hash for ${packageName}`
-        );
-      } else {
-        logger.debug(
-          `github-tags.getTagCommit(): Tag ${tag} not found for ${packageName}`
-        );
-      }
-    } catch (err) {
-      logger.debug(
-        { githubRepo: packageName, err },
-        'Error getting tag commit from GitHub repo'
-      );
-    }
-    return null;
   }
 
   async getCommit(
@@ -91,7 +56,7 @@ export class GithubTagsDatasource extends Datasource {
     newValue?: string
   ): Promise<string | null> {
     return newValue
-      ? this.getTagCommit(registryUrl, repo!, newValue)
+      ? findCommitOfTag(registryUrl, repo!, newValue, this.http)
       : this.getCommit(registryUrl, repo!);
   }
